Improve fusedav tests

+++ tests/saintmode-haproxy.sh
@@ -40,7 +40,7 @@ do
 done
 
 if [ $iters -eq 0 ]; then
-	iters=4
+	iters=1
 fi
 
 pass=0
@@ -50,14 +50,17 @@ if [ $verbose -eq 1 ]; then
     starttime=$(date +%s)
 fi
 
-# This test can only be run on a onebox.
-# It needs to take the haproxies up and down.
+# If run on a onebox, it will take the haproxies up and down.
 # If run outside of a onebox, it would be run on an endpoint,
 # and there is no facility there to take valhalla(yolo)
-# haproxies up and down
-if [ ! -f /etc/systemd/system/haproxy_valhalla21_onebox.service ]; then
-	echo "ERROR: This test needs to be run on a onebox."
-	exit
+# haproxies up and down, so for those tests, other measures
+# need to be taken (e.g. script on each valhallayolo node
+# to stop/restart the haproxy service)
+if [ -f /etc/systemd/system/haproxy_valhalla21_onebox.service ]; then
+	echo "Running on a onebox."
+	onebox=1
+else
+	onebox=0
 fi
 
 # Most tests need to be in the files directory, but this one needs to be
@@ -83,54 +86,57 @@ ha4=haproxy_valhalla.service
 ha4m=8
 D2B=({0..1}{0..1}{0..1}{0..1}{0..1}{0..1}{0..1}{0..1})
 
+uri=$(grep Description /etc/systemd/system/php_fpm_$(pwd | sed s#/srv/bindings/## | sed s#/files##).service | sed s#.*uri=##)
+
 iter=1
 while [ $iter -le $iters ]
 do
 	iter=$((iter + 1))
 
-	up=0
+	if [ $onebox -eq 1 ]; then
+		up=0
 
-	t=$(date +%s); eo=$(($t % 16));
+		t=$(date +%s); eo=$(($t % 16));
 
-	if [ $(($eo & $ha1m)) -eq 0 ]; then
-		systemctl restart $ha1
-		((up+=1))
-	else
-		systemctl stop $ha1
-	fi
+		if [ $(($eo & $ha1m)) -eq 0 ]; then
+			systemctl restart $ha1
+			((up+=1))
+		else
+			systemctl stop $ha1
+		fi
 
-	if [ $(($eo & $ha2m)) -eq 0 ]; then
-		systemctl restart $ha2
-		((up+=2))
-	else
-		systemctl stop $ha2
-	fi
+		if [ $(($eo & $ha2m)) -eq 0 ]; then
+			systemctl restart $ha2
+			((up+=2))
+		else
+			systemctl stop $ha2
+		fi
 
-	if [ $(($eo & $ha3m)) -eq 0 ]; then
-		systemctl restart $ha3
-		((up+=4))
-	else
-		systemctl stop $ha3
-	fi
+		if [ $(($eo & $ha3m)) -eq 0 ]; then
+			systemctl restart $ha3
+			((up+=4))
+		else
+			systemctl stop $ha3
+		fi
 
-	if [ $(($eo & $ha4m)) -eq 0 ]; then
-		systemctl restart $ha4
-		((up+=8))
-	else
-		systemctl stop $ha4
-	fi
+		if [ $(($eo & $ha4m)) -eq 0 ]; then
+			systemctl restart $ha4
+			((up+=8))
+		else
+			systemctl stop $ha4
+		fi
 
-	if [ $verbose -gt 0 ]; then
-		echo "UP: $((10#${D2B[$up]}))"
+		if [ $verbose -gt 0 ]; then
+			echo "UP: $((10#${D2B[$up]}))"
+		fi
 	fi
 
 	for file in $(find files)
 	do 
 		# echo $file
-		res=$(curl -s -I http://dev-panopoly-two.onebox.pantheon.io/sites/default/$file | grep HTTP)
+		res=$(curl -s -I http://$uri/sites/default/$file | grep HTTP)
 		if [ $verbose -gt 0 ]; then
-			#printf "SUCCEED: %s: curl: %s :: %s\n" "$0" "$res" "$file"
-			printf "SUCCEED: %s: %s :: %s\n" "$0" "$file" "$res"
+			printf "SUCCEED: %s: %s : %s :: %s\n" "$0" "$uri" "$file" "$res"
 		fi
 
 		if [[ ! $res =~ '200' && ! $res =~ '301' && ! $res =~ '403' ]]; then
@@ -142,10 +148,12 @@ do
 		sleep 1
 	done
 done
-systemctl restart $ha1
-systemctl restart $ha2
-systemctl restart $ha3
-systemctl restart $ha4
+if [ $onebox -eq 1 ]; then
+	systemctl restart $ha1
+	systemctl restart $ha2
+	systemctl restart $ha3
+	systemctl restart $ha4
+fi
 
 cd files
